requirements.txt
pandas
openpyxl

# embedding model
langchain_community
sentence-transformers

# llm api
langchain_openai